Made MonitorThread an internal class
